rust: Separate Script to Avoid Conflicts
